+++ packages/phorm_generator/lib/builder.dart
@@ -5,12 +5,12 @@ import 'package:path/path.dart' as p;
 import 'package:source_gen/source_gen.dart';
 
 import 'src/model_mixin_generator.dart';
-import 'src/sql_function_generator.dart';
-import 'src/sqlite_schema_generator.dart';
+import 'src/phorm_function_generator.dart';
+import 'src/phorm_schema_generator.dart';
 
 Builder sqlSchemaBuilder(BuilderOptions options) {
   return SharedPartBuilder([
-    SqliteSchemaGenerator(),
+    PhormSchemaGenerator(),
     ModelMixinGenerator(),
   ], 'phorm');
 }
@@ -24,13 +24,13 @@ Builder standaloneSqlSchemaBuilder(BuilderOptions options) {
 
 Builder standaloneSqlFunctionBuilder(BuilderOptions options) {
   return LibraryBuilder(
-    SqlFunctionGenerator(),
+    PhormFunctionGenerator(),
     generatedExtension: '.fn.g.dart',
   );
 }
 
 class _PhormCombinedGenerator extends Generator {
-  final _schemaGen = SqliteSchemaGenerator();
+  final _schemaGen = PhormSchemaGenerator();
   final _mixinGen = ModelMixinGenerator();
 
   @override