+++ packages/phorm_generator/lib/phorm_generator.dart
@@ -1,3 +1,12 @@
 library phorm_generator;
 
-export 'src/sqlite_schema_generator.dart';
+export 'src/generators/function_generator.dart';
+export 'src/generators/mysql/mysql_function_generator.dart';
+export 'src/generators/mysql/mysql_schema_generator.dart';
+export 'src/generators/postgres/postgres_function_generator.dart';
+export 'src/generators/postgres/postgres_schema_generator.dart';
+export 'src/generators/schema_generator.dart';
+export 'src/generators/sqlite/sqlite_function_generator.dart';
+export 'src/generators/sqlite/sqlite_schema_generator.dart';
+export 'src/phorm_function_generator.dart';
+export 'src/phorm_schema_generator.dart';